I ended up cleaning main a bit while at it. Factored out the extra functions to their own Handler class.
