+++ inc/Handler.hpp
@@ -1,0 +1,24 @@
+#pragma once
+#include <iostream>
+#include <vector>
+#include <string>
+#include <queue>
+#include "RecvParser.hpp"
+#include "CommandDispatcher.hpp"
+#include "Client.hpp"
+#include "Server.hpp"
+#include <sys/socket.h>
+#include <sys/types.h>
+#include <netinet/in.h>
+
+extern Server *irc;
+
+class Handler {
+	private:
+		Handler() = delete;
+	public:
+		static void clientWrite(int fd);
+		static void clientDisconnect(int fd);
+		static void acceptClient(int fd);
+
+};
