+++ src/main.cpp
@@ -1,6 +1,5 @@
 #include "Server.hpp"
-#include "RecvParser.hpp"
-#include "CommandDispatcher.hpp"
+#include "Handler.hpp"
 #include <iostream>
 #include <stdexcept>
 #include <sys/socket.h>
@@ -14,49 +13,6 @@ Server *irc;
 
 namespace {
 	volatile std::sig_atomic_t gSigStatus = 0;
-}
-
-void clientWrite(int fd) {
-	int messageLen = 0;
-	string msg;
-	vector<char> buf(BUFSIZ);
-	queue<unique_ptr<Message>> msg_queue;
-	RecvParser	parser(msg_queue);
-	CommandDispatcher	dispatcher;
-
-	messageLen = recv(fd, &buf[0], buf.size(), 0);
-	if (messageLen == -1)
-		throw runtime_error("Failure receiving message");
-	parser.feed(&buf[0], messageLen);
-	while (!msg_queue.empty())
-	{
-		const unique_ptr<Message> &msg = msg_queue.front();
-		dispatcher.dispatch(msg);
-		msg_queue.pop();
-	}
-	send(fd, "Message received\n", 18, 0);
-}
-
-void clientDisconnect(int fd) {
-	cout << "Client with socket" << fd << " disconnected" << endl;
-}
-
-void acceptClient(int socket) {
-	int fd;
-	struct sockaddr_in remote{};
-	socklen_t remoteLen{};
-
-	fd = accept(socket, (struct sockaddr*) &remote, &remoteLen);
-
-	if (fd > 0) {
-		cout << "A client connected with fd nbr " << fd << " connected" << endl;
-		Client client(fd);
-		irc->addClient(client);
-		irc->registerHandler(fd, EPOLLIN, clientWrite);
-		irc->registerHandler(fd, EPOLLRDHUP | EPOLLHUP, clientDisconnect);
-	} else {
-		throw runtime_error("Failed creating a new TCP connection to client");
-	}
 }
 
 int main(int argc, char *argv[]) {
@@ -112,7 +68,8 @@ int main(int argc, char *argv[]) {
 	signal(SIGQUIT, [](int) { gSigStatus = 1; });
 
     irc->addOwnSocket(sock);
-    irc->registerHandler(sock, EPOLLIN, [](int socket) { acceptClient(socket); });
+    irc->registerHandler(sock, EPOLLIN, [](int socket) { Handler::acceptClient(socket); });
+
 	while (!gSigStatus) {
 		try {
 			irc->poll();
